backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
pdfplumber==0.11.0
pdf2image==1.17.0
reportlab==4.2.0
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.0
pytesseract==0.3.10
rouge-score==0.1.2
